+++ src/main/kotlin/blackjack/domain/card/Denomination.kt
@@ -1,4 +1,4 @@
-package blackjack.domain
+package blackjack.domain.card
 
 enum class Denomination(private val score: Score) {
     ACE(Score.of(11)),
