+++ src/main/kotlin/blackjack/domain/deck/Deck.kt
@@ -1,6 +1,6 @@
 package blackjack.domain.deck
 
-import blackjack.domain.Card
+import blackjack.domain.card.Card
 
 class Deck(elements: List<Card>) {
     private val elements = elements.toMutableList()
