+++ src/test/kotlin/blackjack/domain/card/CardTest.kt
@@ -1,5 +1,9 @@
-package blackjack.domain
+package blackjack.domain.card
 
+import blackjack.domain.card.Card
+import blackjack.domain.card.Denomination
+import blackjack.domain.card.Score
+import blackjack.domain.card.Suit
 import org.assertj.core.api.Assertions
 import org.junit.jupiter.api.Test
