+++ src/test/kotlin/blackjack/domain/card/CardsTest.kt
@@ -1,5 +1,10 @@
-package blackjack.domain
+package blackjack.domain.card
 
+import blackjack.domain.card.Card
+import blackjack.domain.card.Cards
+import blackjack.domain.card.Denomination
+import blackjack.domain.card.Score
+import blackjack.domain.card.Suit
 import org.assertj.core.api.Assertions.assertThat
 import org.junit.jupiter.api.Test
